src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
langgraph
python-dotenv
oauthlib
requests-oauthlib
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
openinference-instrumentation-langchain 
opentelemetry-sdk 
opentelemetry-exporter-otlp